+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+2.0.0a3
+-------
+#. Listing fields content and pinned now use a through manager making ordering possible.
+
+2.0.0a2
+-------
+#. Fix case where `resolve()` would fail if site is run from a subpath.
+
 2.0.0a1
 -------
 #. Move to Django 1.6 support. Backwards incompatible.